feat(auth): implementasi login dengan Google OAuth (#29)
* Update : New schema

* Add : New migration

* Update : Sensor controller

* Add : Sensor repository

* Add : Sensor service

* Add : New dependencies

* Del : Log error

* Add : Create prediction function

* Add : Prediction logic

* Update : Get sensor data with prediction result

* Update : New dependencies

* Update : New schema

* Add : New migration

* Update : Index

* Update : Auth controller

* Update : Auth service

* Update : User service

* Add : Passport config

+++ src/user/user.service.js
@@ -32,7 +32,7 @@ const updateUserByIdService = async (userId, data, file) => {
     try {
       const existingImageUrl = existingUser.photo;
 
-      if (existingImageUrl) {
+      if (existingImageUrl && existingImageUrl.publicId) {
         await deleteImageFromCloudinary(existingImageUrl.publicId);
       }
